requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.3
django-cors-headers==4.6.0
django-rest-knox==5.0.2
django-soft-delete==1.0.15
djangorestframework==3.15.2
mysqlclient==2.2.6
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.2
tzdata==2024.2
